+++ com.unity.ml-agents.extensions/Runtime/QTTransform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5fb2e5fd5f1e4c61ba9d510391df7e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
