+++ com.unity.ml-agents.extensions/Runtime/Sensors/HierarchyUtil.cs
@@ -3,55 +3,6 @@ using UnityEngine;
 
 namespace Unity.MLAgents.Extensions.Sensors
 {
-    /// <summary>
-    /// Simple linear transform representation consisting of a rotation and translation.
-    /// When deriving the math, it can be helpful to think of the transform as a 4x4 block matrix:
-    /// <example>
-    /// | R | t |
-    /// ----+----
-    /// | 0 | 1 |
-    /// </example>
-    /// where R is a 3x3 rotation, t is a 3x1 translation, 0 is a 1x3 vector of 0s.
-    /// </summary>
-    public struct QTTransform
-    {
-        public Quaternion Rotation;
-        public Vector3 Translation;
-
-        /// <summary>
-        /// Multiply two transforms.
-        /// <example>
-        /// | R1 | t1 |    | R2 | t2 |   | R1*R2 | R1*t2 + t1 |
-        /// -----+----- *  -----+----- = --------+-------------
-        /// | 0  | 1  |    | 0  | 1  |   |   0   |      1     |
-        /// </example>
-        /// </summary>
-        /// <param name="t1"></param>
-        /// <param name="t2"></param>
-        /// <returns></returns>
-        public static QTTransform operator *(QTTransform t1, QTTransform t2)
-        {
-            var translation = (t1.Rotation * t2.Translation) + t1.Translation;
-            var rotation = t1.Rotation * t2.Rotation;
-            return new QTTransform { Rotation = rotation, Translation = translation };
-        }
-
-        public QTTransform Inverse()
-        {
-            var rotationInverse = Quaternion.Inverse(Rotation);
-            var translationInverse = -(rotationInverse * Translation);
-            return new QTTransform { Rotation = rotationInverse, Translation = translationInverse };
-        }
-
-        public static QTTransform Identity
-        {
-            get { return new QTTransform { Rotation = Quaternion.identity, Translation = Vector3.zero}; }
-        }
-
-        // TODO optimize inv(A)*B?
-
-    }
-
     public abstract class HierarchyUtil
     {
         int[] m_ParentIndices;
@@ -137,79 +88,4 @@ namespace Unity.MLAgents.Extensions.Sensors
         }
     }
 
-    public class RigidBodyHierarchyUtil : HierarchyUtil
-    {
-        Rigidbody[] m_Bodies;
-
-        public void InitTree(Rigidbody rootBody)
-        {
-            // TODO pass root body, walk constraint chain for each body until reach root or parented body
-            var rbs = rootBody.GetComponentsInChildren <Rigidbody>();
-            var joints = rootBody.GetComponentsInChildren <Joint>();
-
-            var parentMap = new Dictionary<Rigidbody, Rigidbody>();
-            foreach (var rb in rbs)
-            {
-                parentMap[rb] = null;
-            }
-
-            foreach (var j in joints)
-            {
-                var parent = j.connectedBody;
-                var child = j.GetComponent<Rigidbody>();
-                parentMap[child] = parent;
-            }
-
-            foreach (var pair in parentMap)
-            {
-                if (pair.Value == null && pair.Key != rootBody)
-                {
-                    Debug.Log($"Found body {pair.Key} with no parent. exiting");
-                    return;
-                }
-            }
-
-            m_Bodies = new Rigidbody[rbs.Length];
-            var parentIndices = new int[rbs.Length];
-            var bodyToIndex = new Dictionary<Rigidbody, int>(rbs.Length);
-
-            m_Bodies[0] = rootBody;
-            parentIndices[0] = -1;
-            bodyToIndex[rootBody] = 0;
-            var index = 1;
-
-            // This is inefficient in the worst case (e.g. a chain)
-            // And might not terminate?
-            while(bodyToIndex.Count != rbs.Length)
-            {
-                foreach (var rb in rbs)
-                {
-                    if (bodyToIndex.ContainsKey(rb))
-                    {
-                        // Already found a place for this one
-                        continue;
-                    }
-
-                    var parent = parentMap[rb];
-                    if (bodyToIndex.ContainsKey(parent))
-                    {
-                        m_Bodies[index] = rb;
-                        parentIndices[index] = bodyToIndex[parent];
-                        bodyToIndex[rb] = index;
-                        index++;
-                    }
-                }
-            }
-
-            SetParentIndices(parentIndices);
-        }
-
-        protected override QTTransform GetTransformAt(int index)
-        {
-            var body = m_Bodies[index];
-            return new QTTransform { Rotation = body.rotation, Translation = body.position };
-        }
-
-
-    }
 }
