+++ com.unity.ml-agents.extensions/Runtime/Sensors/RigidBodyHierarchyUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 393bc7f805ace4702ace3f8daec09630
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
